Improved temporary resource management.

+++ Engine/Source/Core/RenderCore/DX12/SR_CommandList_DX12.h
@@ -75,7 +75,7 @@ public:
 
 	// Copy
 	void CopyResource(SR_Resource* aDstResource, SR_Resource* aSrcResource) override;
-	void CopyBuffer(SR_BufferResource* aDstBuffer, uint32 aDstOffset, SR_BufferResource* aSrcBuffer, uint32 aSrcOffset, uint32 aSize) override;
+	void CopyBuffer(SR_BufferResource* aDstBuffer, uint64 aDstOffset, SR_BufferResource* aSrcBuffer, uint64 aSrcOffset, uint32 aSize) override;
 	void CopyTexture(SR_TextureResource* aDstTexture, SR_TextureResource* aSrcResource) override;
 
 	void UpdateBuffer(SR_BufferResource* aDstBuffer, uint32 aDstOffset, const void* aData, uint32 aSize) override;
